requirements.txt
# InfraMIND v3 — Dependencies
# Core
numpy>=1.24.0
scipy>=1.10.0
simpy>=4.1.1
pyyaml>=6.0
dataclasses-json>=0.6.0

# Optimization
torch>=2.0.0
gpytorch>=1.11
botorch>=0.9.0

# Structure Learning
scikit-learn>=1.3.0

# Visualization & Plotting
matplotlib>=3.7.0
seaborn>=0.12.0

# Testing
pytest>=7.4.0

# Utilities
tqdm>=4.65.0
tabulate>=0.9.0
